backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
alembic==1.13.1
